backend/module_payments/requirements.txt
# /module_payments/requirements.txt
fastapi
uvicorn
midtransclient
python-dotenv
sqlalchemy # Untuk interaksi DB (contoh)
psycopg2-binary # Driver Postgres